apply single resposible princepal

+++ src/main/java/com/service/VideoService.java
@@ -1,9 +1,13 @@
 package com.service;
 
+import com.exceptions.ResourceNotFoundException;
+import com.model.Section;
 import com.model.Video;
 import com.repository.VideoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 @Service
 public class VideoService {
@@ -14,12 +18,31 @@ public class VideoService {
     public  Video save(Video video) {
           return videoRepository.save(video);
     }
-
-    public Video getVideoById(Long id) {
-        return videoRepository.getOne(id);
+    public Section addVideosToSection(Section section, List<Video> videos) {
+        for (Video video:videos
+        ) {
+            section.addVideos(video);
+           videoRepository.save(video);
+        }
+        return section;
     }
 
-    public void deleteVideo(Video video) {
+    public Video updateVideo(Video video) {
+        Video video1 = videoRepository.getOne(video.getId());
+        if (video1==null){
+            throw new ResourceNotFoundException("Video with id "+video.getId()+" not exist");
+
+        }
+        video1 = videoRepository.save(video);
+        return video1;
+    }
+
+    public void deleteVideo(Long videoId) {
+        Video video = videoRepository.getOne(videoId);
+        if (video == null){
+            throw new ResourceNotFoundException("Video with id "+videoId+" not exist");
+        }
         videoRepository.delete(video);
+
     }
 }
